Move tests for utils
